message spam update

+++ admin_products.php
@@ -25,17 +25,14 @@ if(isset($_POST['add_product'])){
     if(mysqli_num_rows($select_product_name)>0){
         $message[]='product name already added';
 
-
-
-
     }else{
-        $add_product_query=mysqli_query($conn, "INSERT INTO `products`(name,price,image) VALUES('$name','$price','$image')")or die('query failed');
+        $add_product_query=mysqli_query($conn, "INSERT INTO `products`(name, price, image) VALUES('$name', '$price', '$image')")or die('query failed');
 
         if($add_product_query){
             if($image_size>2000000){
                 $message[]='image size is too large';
             }else{
-            move_uploaded_file($image_tmp_name,$image_folder);
+            move_uploaded_file($image_tmp_name, $image_folder);
             $message[]='product added succesfully';
             }
         }else{
@@ -105,7 +102,7 @@ if(isset($_POST['add_product'])){
             <div class="name"><?php echo $fetch_products['name']; ?></div>
             <div class="price"><?php echo $fetch_products['price']; ?>/-</div>
             <a href="admin_products.php?update=<?php echo $fetch_products['id']; ?>"class="option-btn">Update</a>
-            <a href="admin_products.php?delete=<?php echo $fetch_products['id']; ?>"class="delete-btn">Delete</a>
+            <a href="admin_products.php?delete=<?php echo $fetch_products['id']; ?>"class="delete-btn" onclick="return confirm('delete this product')">Delete</a>
 
 
         </div>
@@ -128,11 +125,6 @@ if(isset($_POST['add_product'])){
         <!--custom admin js file link -->
 
         <script src="js/admin_script.js"></script>
-
-
-
-
-
 
 
     </body>
